Fix musl build
it seems we need to explicitly include `stdio.h` for
vaccelrt to built correctly using musl.

```
export CC="musl-gcc"
mkdir build && cd build
cmake -DCMAKE_EXE_LINKER_FLAGS="-static -Os" ..
make
```

```
$ musl-ldd src/libvaccel.so
	musl-ldd (0x7fef4be96000)
	libc.so => musl-ldd (0x7fef4be96000)
```

Also, we cleanup some of the example code, removing redundant functions.
