+++ CMakeLists.txt
@@ -10,6 +10,7 @@ option(BUILD_PLUGIN_EXEC "Build the exec plugin" OFF)
 option(BUILD_PLUGIN_NOOP "Build the no-op debugging plugin" OFF)
 option(BUILD_EXAMPLES "Build the examples" OFF)
 option(ENABLE_TESTS "Enable testing" OFF)
+option(MUSL_BUILD "Link against MUSL" OFF)
 
 # Export json with compile commands
 SET(CMAKE_EXPORT_COMPILE_COMMANDS ON)
