+++ examples/common.h
@@ -1,0 +1,61 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+static int read_file(const char *filename, char **img, size_t *img_size)
+{
+	int fd;
+
+	fd = open(filename, O_RDONLY);
+	if (fd < 0) {
+		perror("open: ");
+		return 1;
+	}
+
+	struct stat info;
+	fstat(fd, &info);
+	fprintf(stdout, "Image size: %luB\n", info.st_size);
+
+	char *buf = malloc(info.st_size);
+	if (!buf) {
+		fprintf(stderr, "Could not allocate memory for image\n");
+		goto close_file;
+	}
+
+	long bytes = 0;
+	do {
+		int ret = read(fd, buf, info.st_size);
+		if (ret < 0) {
+			perror("Error while reading image: ");
+			goto free_buff;
+		}
+		bytes += ret;
+	} while (bytes < info.st_size);
+
+	if (bytes < info.st_size) {
+		fprintf(stderr, "Could not read image\n");
+		goto free_buff;
+	}
+
+	*img = buf;
+	*img_size = info.st_size;
+	close(fd);
+
+	return 0;
+
+free_buff:
+	free(buf);
+close_file:
+	close(fd);
+	return 1;
+}
