+++ src/CMakeLists.txt
@@ -5,6 +5,7 @@ file(GLOB_RECURSE sources *.c)
 file(GLOB_RECURSE headers *.h)
 
 add_library(vaccel SHARED ${headers} ${sources})
+
 target_compile_options(vaccel PUBLIC -Wall -Wextra -Werror -pthread)
 target_include_directories(vaccel PRIVATE ${CMAKE_SOURCE_DIR}/src)
 target_include_directories(vaccel PRIVATE ${CMAKE_SOURCE_DIR}/src/include)
@@ -12,6 +13,9 @@ target_include_directories(vaccel PRIVATE ${CMAKE_BINARY_DIR})
 
 target_include_directories(vaccel PRIVATE ${CMAKE_SOURCE_DIR}/src/include)
 set_property(TARGET vaccel PROPERTY LINK_FLAGS "-pthread")
+IF(MUSL_BUILD)
+set_property(TARGET vaccel PROPERTY LINK_FLAGS "-Os -pthread")
+ENDIF()
 set_property(TARGET vaccel PROPERTY C_STANDARD 11)
 target_link_libraries(vaccel slog)
